# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Openpage API",
        "version": "1.0.0"
    },
    "basePath": "/openpage-backend/public",
    "paths": {
...
        "/media": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Returns list of media",
                "description": "Returns media data",
                "operationId": "media",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "search_key",
                        "in": "query",
                        "description": "Search based on remark and tags",
                        "type": "string"
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "this is offset or skip the records",
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of records to be retrieved ",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Media"
                ],
                "summary": "Update media file data",
                "description": "Update media file in the system",
                "operationId": "update_media",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "mid",
                        "in": "query",
                        "description": "ID of the media that needs to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "usage[]",
                        "in": "query",
                        "description": "The page ids array field.",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "remark",
                        "in": "query",
                        "description": "Remark of the media",
                        "type": "string"
                    },
                    {
                        "name": "tag[]",
                        "in": "query",
                        "description": "tags for the media.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid data supplied"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Media"
                ],
                "summary": "Creating/Storing new media file",
                "description": "Stores media file in the system",
                "operationId": "create_media",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type of media file to be uploaded",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "extension",
                        "in": "query",
                        "description": "Extension of media file",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "media_file",
                        "in": "formData",
                        "description": "file to upload",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "usage[]",
                        "in": "query",
                        "description": "The page ids array field.",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "remark",
                        "in": "query",
                        "description": "Remark of the media",
                        "type": "string"
                    },
                    {
                        "name": "tag[]",
                        "in": "query",
                        "description": "tags for the media.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid data supplied"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Media"
                ],
                "summary": "delete media file data",
                "description": "Delete media file in the system",
                "operationId": "delete_media",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "mid",
                        "in": "query",
                        "description": "ID of the media that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid data supplied"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/media/{mid}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Returns media data",
                "description": "Returns media data",
                "operationId": "media",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "mid",
                        "in": "path",
                        "description": "ID of the media that needs to be displayed",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid media id"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    },
...
}